+++ CHANGELOG.md
@@ -8,6 +8,9 @@ to a section with the version name.
 
 ## Unreleased Changes
 
+* Track which `Sales Invoice Additional Fields` is latest in case of multiple submissions for the same invoice  due to
+  rejection
+
 ## 0.20.2
 
 * Validate that sales invoice has tax rate in Sales Taxes and Charges Table in Phase 1 and Phase 2.